requirements.python-services.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
torch>=2.0.1
transformers>=4.30.0
sentence-transformers>=2.2.2
openai>=1.3.0
google-generativeai>=0.5.0
geopy>=2.4.0
pydeck>=0.8.0
pandas>=2.0.0
plotly>=5.15.0
altair>=5.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
PyYAML>=6.0
scikit-learn>=1.3.0
aiosqlite
sqlalchemy[asyncio]>=2.0.0
sqlmodel
alembic
asyncpg>=0.27.0
requests>=2.31.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx
python-multipart